+++ lib/src/models/yust_file_scan.g.dart
@@ -1,0 +1,15 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+part of 'yust_file_scan.dart';
+
+// **************************************************************************
+// JsonSerializableGenerator
+// **************************************************************************
+
+Map<String, dynamic> _$YustFileScanToJson(YustFileScan instance) =>
+    <String, dynamic>{
+      'status': instance.status.toJson(),
+      'signature': instance.signature,
+      'reason': instance.reason?.toJson(),
+      'scannedAt': instance.scannedAt?.toIso8601String(),
+    };