+++ lib/src/models/yust_image.g.dart
@@ -6,28 +6,36 @@ part of 'yust_image.dart';
 // JsonSerializableGenerator
 // **************************************************************************
 
-YustImage _$YustImageFromJson(Map json) => YustImage(
-  name: json['name'] as String?,
-  modifiedAt: json['modifiedAt'] == null
-      ? null
-      : DateTime.parse(json['modifiedAt'] as String),
-  url: json['url'] as String?,
-  hash: json['hash'] as String? ?? '',
-  createdAt: json['createdAt'] == null
-      ? null
-      : DateTime.parse(json['createdAt'] as String),
-  path: json['path'] as String?,
-  thumbnails: (json['thumbnails'] as Map?)?.map(
-    (k, e) =>
-        MapEntry($enumDecode(_$YustFileThumbnailSizeEnumMap, k), e as String),
-  ),
-  favorite: json['favorite'] as bool? ?? false,
-  location: json['location'] == null
-      ? null
-      : YustGeoLocation.fromJson(
-          Map<String, dynamic>.from(json['location'] as Map),
+YustImage _$YustImageFromJson(Map json) =>
+    YustImage(
+        name: json['name'] as String?,
+        modifiedAt: json['modifiedAt'] == null
+            ? null
+            : DateTime.parse(json['modifiedAt'] as String),
+        url: json['url'] as String?,
+        hash: json['hash'] as String? ?? '',
+        createdAt: json['createdAt'] == null
+            ? null
+            : DateTime.parse(json['createdAt'] as String),
+        path: json['path'] as String?,
+        thumbnails: (json['thumbnails'] as Map?)?.map(
+          (k, e) => MapEntry(
+            $enumDecode(_$YustFileThumbnailSizeEnumMap, k),
+            e as String,
+          ),
         ),
-);
+        favorite: json['favorite'] as bool? ?? false,
+        location: json['location'] == null
+            ? null
+            : YustGeoLocation.fromJson(
+                Map<String, dynamic>.from(json['location'] as Map),
+              ),
+      )
+      ..scan = json['scan'] == null
+          ? null
+          : YustFileScan.fromJson(
+              Map<String, dynamic>.from(json['scan'] as Map),
+            );
 
 Map<String, dynamic> _$YustImageToJson(YustImage instance) => <String, dynamic>{
   'name': instance.name,
@@ -38,6 +46,7 @@ Map<String, dynamic> _$YustImageToJson(YustImage instance) => <String, dynamic>{
   'path': instance.path,
   'favorite': instance.favorite,
   'thumbnails': instance.thumbnails?.map((k, e) => MapEntry(k.toJson(), e)),
+  'scan': instance.scan?.toJson(),
   'location': instance.location?.toJson(),
 };